Refactor fuzzy search functionality across Extensions and Library components
- Moved the fuzzy search function to a shared utility, allowing for consistent usage across ExtensionsScreenModel and LibraryItem.
- Removed redundant fuzzy search implementations from individual components, improving code maintainability and readability.
- Updated imports to utilize the new shared function, ensuring all components benefit from the same search logic.

These changes streamline the search functionality and enhance the overall code structure.
