+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/extension/ExtensionsScreenModel.kt
@@ -33,6 +33,7 @@ import tachiyomi.i18n.MR
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import kotlin.time.Duration.Companion.seconds
+import eu.kanade.tachiyomi.util.lang.containsFuzzy
 
 class ExtensionsScreenModel(
     preferences: SourcePreferences = Injekt.get(),
@@ -56,27 +57,6 @@ class ExtensionsScreenModel(
                 query.split(",").any { _input ->
                     val input = _input.trim()
                     if (input.isEmpty()) return@any false
-
-                    // Fuzzy search function that checks if characters appear in sequence
-                    fun String.containsFuzzy(other: String, ignoreCase: Boolean = true): Boolean {
-                        if (other.isEmpty()) return true
-                        if (this.isEmpty()) return false
-
-                        val sourceText = if (ignoreCase) this.lowercase() else this
-                        val searchText = if (ignoreCase) other.lowercase() else other
-
-                        var sourceIndex = 0
-                        var searchIndex = 0
-
-                        while (sourceIndex < sourceText.length && searchIndex < searchText.length) {
-                            if (sourceText[sourceIndex] == searchText[searchIndex]) {
-                                searchIndex++
-                            }
-                            sourceIndex++
-                        }
-
-                        return searchIndex == searchText.length
-                    }
 
                     when (extension) {
                         is Extension.Available -> {
