+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryItem.kt
@@ -1,6 +1,8 @@
 package eu.kanade.tachiyomi.ui.library
 
+import android.content.Context
 import eu.kanade.tachiyomi.source.getNameForMangaInfo
+import eu.kanade.tachiyomi.util.lang.containsFuzzy
 import tachiyomi.domain.library.model.LibraryManga
 import tachiyomi.domain.source.service.SourceManager
 import uy.kohesive.injekt.Injekt
@@ -22,27 +24,6 @@ data class LibraryItem(
      */
     fun matches(constraint: String): Boolean {
         val sourceName by lazy { sourceManager.getOrStub(libraryManga.manga.source).getNameForMangaInfo(null) }
-
-        // Fuzzy search function that checks if characters appear in sequence
-        fun String.containsFuzzy(other: String, ignoreCase: Boolean = true): Boolean {
-            if (other.isEmpty()) return true
-            if (this.isEmpty()) return false
-
-            val sourceText = if (ignoreCase) this.lowercase() else this
-            val searchText = if (ignoreCase) other.lowercase() else other
-
-            var sourceIndex = 0
-            var searchIndex = 0
-
-            while (sourceIndex < sourceText.length && searchIndex < searchText.length) {
-                if (sourceText[sourceIndex] == searchText[searchIndex]) {
-                    searchIndex++
-                }
-                sourceIndex++
-            }
-
-            return searchIndex == searchText.length
-        }
 
         return libraryManga.manga.title.containsFuzzy(constraint) ||
             (libraryManga.manga.author?.containsFuzzy(constraint) ?: false) ||
