+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryScreenModel.kt
@@ -121,6 +121,7 @@ import tachiyomi.source.local.isLocal
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import kotlin.random.Random
+import eu.kanade.tachiyomi.util.lang.containsFuzzy
 
 /**
  * Typealias for the library manga, using the category as keys, and list of manga as values.
@@ -1011,27 +1012,6 @@ class LibraryScreenModel(
         val sourceIdString = manga.source.takeUnless { it == LocalSource.ID }?.toString()
         val genre = if (checkGenre) manga.genre.orEmpty() else emptyList()
         val context = Injekt.get<Application>()
-
-        // Fuzzy search function that checks if characters appear in sequence
-        fun String.containsFuzzy(other: String, ignoreCase: Boolean = true): Boolean {
-            if (other.isEmpty()) return true
-            if (this.isEmpty()) return false
-
-            val sourceText = if (ignoreCase) this.lowercase() else this
-            val searchText = if (ignoreCase) other.lowercase() else other
-
-            var sourceIndex = 0
-            var searchIndex = 0
-
-            while (sourceIndex < sourceText.length && searchIndex < searchText.length) {
-                if (sourceText[sourceIndex] == searchText[searchIndex]) {
-                    searchIndex++
-                }
-                sourceIndex++
-            }
-
-            return searchIndex == searchText.length
-        }
 
         return queries.all { queryComponent ->
             when (queryComponent.excluded) {
